+++ src/vctrs.h
@@ -263,6 +263,7 @@ enum name_repair_arg {
   name_repair_check_unique
 };
 
+const char* name_repair_arg_as_c_string(enum name_repair_arg arg);
 enum name_repair_arg validate_name_repair(SEXP arg);
 SEXP vec_as_names(SEXP names, enum name_repair_arg type, bool quiet);
 bool is_unique_names(SEXP names);