+++ NEWS.md
@@ -1,8 +1,10 @@
 # vctrs 0.1.0.9000
 
+* `vec_cbind()` gains a `.name_repair` argument.
+
 * `vec_dims()` has been replaced by `vec_dim_n()`. This name more clearly
   indicates the purpose of the function.
-  
+
 * New `stop_incompatible_size()` to signal a failure due to mismatched sizes.
 
 * `vec_empty()` has been renamed to `vec_is_empty()`.